+++ examples/render.rs
@@ -32,7 +32,7 @@ pub fn main() {
 
     new_base_widget
         .get_callbacks()
-        .on_mouse_entered(|x, _widgets| {
+        .on_mouse_entered(|x, _widgets, _layouts| {
             x.get_config()
                 .set_color(CONFIG_COLOR_BASE, Color::RGB(255, 0, 0));
             x.get_config().set_invalidate(true);
@@ -46,7 +46,7 @@ pub fn main() {
 
     new_base_widget
         .get_callbacks()
-        .on_mouse_exited(|x, _widgets| {
+        .on_mouse_exited(|x, _widgets, _layouts| {
             x.get_config()
                 .set_color(CONFIG_COLOR_BASE, Color::RGB(255, 255, 255));
             x.get_config().set_invalidate(true);
@@ -60,19 +60,19 @@ pub fn main() {
 
     new_base_widget
         .get_callbacks()
-        .on_mouse_moved(|_widget, _widgets, points| {
+        .on_mouse_moved(|_widget, _widgets, _layouts, points| {
             eprintln!("Mouse Moved: {:?}", points);
         });
 
     new_base_widget
         .get_callbacks()
-        .on_mouse_scrolled(|_widget, _widgets, points| {
+        .on_mouse_scrolled(|_widget, _widgets, _layouts, points| {
             eprintln!("Mouse Scrolled: {:?}", points);
         });
 
     new_base_widget
         .get_callbacks()
-        .on_mouse_clicked(|_widget, _widgets, button, clicks, state| {
+        .on_mouse_clicked(|_widget, _widgets, _layouts, button, clicks, state| {
             eprintln!(
                 "Mouse Clicked: button={} clicks={} state={}",
                 button, clicks, state