+++ src/render/engine.rs
@@ -101,6 +101,7 @@ impl Engine {
                             mouse_btn as u8,
                             clicks,
                             true,
+                            self.layout_cache.get_layout_cache(),
                         );
                     }
 
@@ -108,7 +109,9 @@ impl Engine {
                         mouse_btn, clicks, ..
                     } => {
                         self.widget_cache
-                            .button_clicked(-1, mouse_btn as u8, clicks, false);
+                            .button_clicked(-1, mouse_btn as u8, clicks, false,
+                                            self.layout_cache.get_layout_cache(),
+                            );
                     }
 
                     Event::MouseMotion { x, y, .. } => {
@@ -117,17 +120,25 @@ impl Engine {
                         self.current_widget_id = self.widget_cache.find_widget(x, y);
 
                         if cur_widget_id != self.current_widget_id {
-                            self.widget_cache.mouse_exited(cur_widget_id);
-                            self.widget_cache.mouse_entered(self.current_widget_id);
+                            self.widget_cache.mouse_exited(cur_widget_id,
+                                                           self.layout_cache.get_layout_cache(),
+                            );
+                            self.widget_cache.mouse_entered(self.current_widget_id,
+                                                            self.layout_cache.get_layout_cache(),
+                            );
                         }
 
                         self.widget_cache
-                            .mouse_moved(self.current_widget_id, vec![x, y]);
+                            .mouse_moved(self.current_widget_id, vec![x, y],
+                                         self.layout_cache.get_layout_cache(),
+                            );
                     }
 
                     Event::MouseWheel { x, y, .. } => {
                         self.widget_cache
-                            .mouse_scrolled(self.current_widget_id, vec![x, y]);
+                            .mouse_scrolled(self.current_widget_id, vec![x, y],
+                                            self.layout_cache.get_layout_cache(),
+                            );
                     }
 
                     Event::Quit { .. } => {
@@ -136,12 +147,14 @@ impl Engine {
 
                     remaining_event => {
                         self.widget_cache
-                            .other_event(self.current_widget_id, remaining_event);
+                            .other_event(self.current_widget_id, remaining_event,
+                                         self.layout_cache.get_layout_cache(),
+                            );
                     }
                 }
             }
 
-            self.widget_cache.tick();
+            self.widget_cache.tick(self.layout_cache.get_layout_cache());
             self.layout_cache
                 .do_layout(self.widget_cache.borrow_cache());
             self.widget_cache.draw_loop(&mut canvas);