+++ src/render/widget_cache.rs
@@ -22,6 +22,7 @@ use sdl2::pixels::Color;
 use sdl2::rect::Rect;
 use sdl2::render::Canvas;
 use sdl2::video::Window;
+use crate::render::layout_cache::LayoutContainer;
 
 /// This is a container that stores information about a `Widget` that will be drawn on the screen.
 /// It stores the `Widget` object, the actual point of origin inside the `Window` (as a `Vec<i32>`
@@ -176,12 +177,13 @@ impl WidgetCache {
     /// to `false`, it indicates that the mouse button was released.  When setting the button state
     /// to `widget_id == -1`, the button click message will be sent to _all_ `Widget`s, so use
     /// `widget_id == - 1` with care.
-    pub fn button_clicked(&mut self, widget_id: i32, button: u8, clicks: u8, state: bool) {
+    pub fn button_clicked(&mut self, widget_id: i32, button: u8, clicks: u8, state: bool, cache: &[LayoutContainer]) {
         if widget_id == -1 {
             for i in 0..self.cache.len() {
                 if !self.is_hidden(i as i32) && self.is_enabled(i as i32) {
                     self.cache[i as usize].widget.borrow_mut().button_clicked(
                         &self.cache,
+                        cache,
                         button,
                         clicks,
                         state,
@@ -192,71 +194,71 @@ impl WidgetCache {
             self.cache[widget_id as usize]
                 .widget
                 .borrow_mut()
-                .button_clicked(&self.cache, button, clicks, state);
+                .button_clicked(&self.cache, cache, button, clicks, state);
         }
     }
 
     /// This function calls the `mouse_moved` callback for the `Widget` specified by `widget_id`.
-    pub fn mouse_moved(&mut self, widget_id: i32, points: Vec<i32>) {
+    pub fn mouse_moved(&mut self, widget_id: i32, points: Vec<i32>, cache: &[LayoutContainer]) {
         if !self.is_hidden(widget_id) && self.is_enabled(widget_id) {
             self.cache[widget_id as usize]
                 .widget
                 .borrow_mut()
-                .mouse_moved(&self.cache, points);
+                .mouse_moved(&self.cache, cache, points);
         }
     }
 
     /// This function calls the `mouse_scrolled` callback for the `Widget` specified by `widget_id`.
-    pub fn mouse_scrolled(&mut self, widget_id: i32, points: Vec<i32>) {
+    pub fn mouse_scrolled(&mut self, widget_id: i32, points: Vec<i32>, cache: &[LayoutContainer]) {
         if !self.is_hidden(widget_id) && self.is_enabled(widget_id) {
             self.cache[widget_id as usize]
                 .widget
                 .borrow_mut()
-                .mouse_scrolled(&self.cache, points);
+                .mouse_scrolled(&self.cache, cache, points);
         }
     }
 
     /// This function calls the `mouse_exited` callback for the `Widget` specified by `widget_id`.
-    pub fn mouse_exited(&mut self, widget_id: i32) {
+    pub fn mouse_exited(&mut self, widget_id: i32, cache: &[LayoutContainer]) {
         if !self.is_hidden(widget_id) && self.is_enabled(widget_id) {
             self.cache[widget_id as usize]
                 .widget
                 .borrow_mut()
-                .mouse_exited(&self.cache);
+                .mouse_exited(&self.cache, cache);
         }
     }
 
     /// This function calls the `mouse_entered` callback for the `Widget` specified by `widget_id`.
-    pub fn mouse_entered(&mut self, widget_id: i32) {
+    pub fn mouse_entered(&mut self, widget_id: i32, cache: &[LayoutContainer]) {
         if !self.is_hidden(widget_id) && self.is_enabled(widget_id) {
             self.cache[widget_id as usize]
                 .widget
                 .borrow_mut()
-                .mouse_entered(&self.cache);
+                .mouse_entered(&self.cache, cache);
         }
     }
 
     /// This function calls the `tick` method on all registered `Widget`s in the cache.  The purpose
     /// for the `tick` is to indicate that a drawing loop is about to occur, and the `Widget` can
     /// update itself as necessary beforehand.
-    pub fn tick(&mut self) {
+    pub fn tick(&mut self, _cache: &[LayoutContainer]) {
         let cache_size = self.cache.len();
 
         for i in 0..cache_size {
             if !self.is_hidden(i as i32) {
-                self.cache[i].widget.borrow_mut().tick(&self.cache);
+                self.cache[i].widget.borrow_mut().tick(&self.cache, _cache);
             }
         }
     }
 
     /// This function sends all other un-handled events from SDL2 to the currently highlighted
     /// `Widget`.
-    pub fn other_event(&mut self, widget_id: i32, event: Event) {
+    pub fn other_event(&mut self, widget_id: i32, event: Event, cache: &[LayoutContainer]) {
         if !self.is_hidden(widget_id) && self.is_enabled(widget_id) {
             self.cache[widget_id as usize]
                 .widget
                 .borrow_mut()
-                .other_event(&self.cache, event);
+                .other_event(&self.cache, cache, event);
         }
     }