+++ src/widgets/progress_widget.rs
@@ -16,6 +16,7 @@
 use crate::render::callbacks::CallbackRegistry;
 use crate::render::widget::*;
 use crate::render::widget_cache::WidgetContainer;
+use crate::render::layout_cache::LayoutContainer;
 use crate::render::widget_config::*;
 use crate::render::Points;